+++ lib/holder/main/componenets/positioned_button.dart
@@ -1,21 +1,1 @@
-import 'package:flutter/cupertino.dart';
-
-import '../../../components/bottom_button.dart';
-
-class PositionedButton extends StatelessWidget {
-  String text;
-
-  PositionedButton({
-    required this.text,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return Positioned(
-      left: 10,
-      right: 10,
-      bottom: 60,
-      child: BottomButton(text: text),
-    );
-  }
-}
+// TODO Implement this library.
