+++ lib/holder/main/components/home_auto_ad_slider.dart
